Last Update

+++ app.py
@@ -7,6 +7,9 @@ SUPABASE_URL = "https://etxyomkjynglfwmopytg.supabase.co"
 SUPABASE_KEY = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImV0eHlvbWtqeW5nbGZ3bW9weXRnIiwicm9sZSI6ImFub24iLCJpYXQiOjE3NTIwNDAzODcsImV4cCI6MjA2NzYxNjM4N30.ZWlhHLvpfLJ5E-oivh-BrdM01fpUzJOigdwgqi7__4E"
 supabase: Client = create_client(SUPABASE_URL, SUPABASE_KEY)
 
+# Admin code
+kodeadmin = "2344"  # Ganti dengan kode admin yang diinginkan
+
 app = Flask(__name__)
 
 import random, string
@@ -14,11 +17,16 @@ import random, string
 def generate_code(length=6):
     return ''.join(random.choices(string.ascii_letters + string.digits, k=length))
 
-@app.route('/', methods=['GET', 'POST'])
+@app.route('/posturl', methods=['GET', 'POST'])
 def home():
     if request.method == 'POST':
         long_url = request.form['url']
         custom_code = request.form['custom_code'].strip()
+        admin_code = request.form['admincode'].strip()
+
+        # Validasi admin code
+        if admin_code != kodeadmin:
+            return "❌ Admin code salah. Akses ditolak.", 403
 
         # Jika user isi custom code
         if custom_code:
@@ -37,6 +45,7 @@ def home():
         return render_template('result.html', short_url=short_url)
 
     return render_template('form.html')
+
 
 @app.route('/<code>')
 def redirect_to_url(code):
@@ -61,6 +70,11 @@ def edit(code):
     if request.method == 'POST':
         new_url = request.form['new_url'].strip()
         new_code = request.form['new_code'].strip()
+        admin_code = request.form['admincode'].strip()
+
+        # Validasi admin code
+        if admin_code != kodeadmin:
+            return "❌ Admin code salah. Akses ditolak.", 403
 
         # Jika kode baru sudah digunakan dan beda dari yang sekarang
         if new_code != code:
@@ -78,11 +92,17 @@ def edit(code):
 # Delete
 @app.route('/delete/<code>', methods=['GET', 'POST'])
 def delete(code):
+
     result = supabase.table("urls").select("*").eq("code", code).execute()
+
     if not result.data:
         return "Shortcode tidak ditemukan", 404
 
     if request.method == 'POST':
+        admin_code = request.form.get('admincode', '').strip()
+        # Validasi admin code
+        if admin_code != kodeadmin:
+            return "❌ Admin code salah. Akses ditolak.", 403
         supabase.table("urls").delete().eq("code", code).execute()
         return redirect('/admin')
